requirements.txt
graphics.py
screeninfo